refactor(agents): give shared box plumbing a home that is not claude.ts
codex.ts, opencode.ts, seed.ts and shell-session.ts all imported
buildTermSafeTmuxExec / buildTmuxSessionArgs / CONTAINER_USER from claude.ts,
which made Claude the de facto shared module for the whole package -- including
for agentbox shell, which has no agent at all.

Harmless while the three are siblings in one directory, fatal once each agent
becomes its own package: every agent package would depend on
@agentbox/agent-claude to get a tmux flag, and 'an agent is a package' would be
false on arrival. Nothing here was ever Claude-specific.

shared.ts is deliberately dependency-free, and a test asserts both halves: no
agent module imports another, and shared.ts imports nothing. Public exports are
unchanged -- the barrel just sources them from their new home.
